AccessControl: tidy scope resolution
